Explode searchdialog.py in a submodule
